+++ src/components/FlawForm/FlawForm.vue
@@ -28,7 +28,13 @@ import LabelStatic from '@/widgets/LabelStatic/LabelStatic.vue';
 import LabelSelect from '@/widgets/LabelSelect/LabelSelect.vue';
 import LabelTagsInput from '@/widgets/LabelTagsInput/LabelTagsInput.vue';
 import LabelEditable from '@/widgets/LabelEditable/LabelEditable.vue';
-import { type ZodFlawType, descriptionRequiredStates, flawImpactEnum, flawSources } from '@/types/zodFlaw';
+import {
+  type ZodFlawType,
+  MajorIncidentStateEnumWithBlank,
+  descriptionRequiredStates,
+  flawImpactEnum,
+  flawSources,
+} from '@/types/zodFlaw';
 import { useDraftFlawStore } from '@/stores/DraftFlawStore';
 import { deepCopyFromRaw } from '@/utils/helpers';
 import { allowedSources } from '@/constants/';
@@ -61,7 +67,6 @@ const {
   errors,
   flaw,
   flawAcknowledgments,
-  flawIncidentStates,
   flawReferences,
   internalComments,
   internalCommentsAvailable,
@@ -295,7 +300,7 @@ const createdDate = computed(() => {
             <LabelSelect
               v-model="flaw.major_incident_state"
               label="Incident State"
-              :options="flawIncidentStates"
+              :options="MajorIncidentStateEnumWithBlank"
               :error="errors.major_incident_state"
             />
             <LabelEditable